requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dbus-python==1.4.0
distro==1.9.0
Flask==3.1.2
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-genai==1.36.0
graphviz==0.21
groq==0.31.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
logging==0.4.9.6
MarkupSafe==3.0.2
matplotlib==3.10.6
narwhals==2.5.0
numpy==2.3.3
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
plotly==6.3.0
plyer==2.1.0
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
schedule==1.2.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
streamlit==1.49.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
